refactor: use return type of clients factory
