--- apps/gql/clients/types.ts
@@ -1,5 +1,0 @@
-import { type PrismaClient } from "@kampus/prisma";
-
-export interface Clients {
-  prisma: PrismaClient;
-}
